Remove redundant parts in filenames
