+++ README.md
@@ -440,7 +440,7 @@ Have a question? Found a bug? Missing a specific feature? Ran into a problem? Fe
 Every run is initialized by [run.py](run.py) file. All PyTorch Lightning modules are dynamically instantiated from module paths specified in config. Example model config:
 
 ```yaml
-_target_: src.models.mnist_model.MNISTLitModel
+_target_: src.models.mnist.MNISTLitModel
 input_size: 784
 lin1_size: 256
 lin2_size: 256
@@ -475,8 +475,8 @@ It also specifies everything that shouldn't be managed by experiment configurati
 # specify here default training configuration
 defaults:
   - trainer: default.yaml
-  - model: mnist_model.yaml
-  - datamodule: mnist_datamodule.yaml
+  - model: mnist.yaml
+  - datamodule: mnist.yaml
   - callbacks: default.yaml # set this to null if you don't want to use callbacks
   - logger: null # set logger here or use command line (e.g. `python run.py logger=wandb`)
 
@@ -520,8 +520,8 @@ Experiment configurations allow you to overwrite parameters from main project co
 defaults:
   - override /mode: exp.yaml
   - override /trainer: default.yaml
-  - override /model: mnist_model.yaml
-  - override /datamodule: mnist_datamodule.yaml
+  - override /model: mnist.yaml
+  - override /datamodule: mnist.yaml
   - override /callbacks: default.yaml
   - override /logger: null
 
@@ -581,7 +581,7 @@ trainer:
   gradient_clip_val: 0.5
 
 model:
-  _target_: src.models.mnist_model.MNISTLitModel
+  _target_: src.models.mnist.MNISTLitModel
   lr: 0.001
   weight_decay: 0.00005
   input_size: 784
@@ -591,7 +591,7 @@ model:
   output_size: 10
 
 datamodule:
-  _target_: src.datamodules.mnist_datamodule.MNISTDataModule
+  _target_: src.datamodules.mnist.MNISTDataModule
   data_dir: ${data_dir}
   train_val_test_split: [55_000, 5_000, 10_000]
   batch_size: 64
@@ -644,8 +644,8 @@ hydra:
 
 ### Workflow
 
-1. Write your PyTorch Lightning model (see [mnist_model.py](src/models/mnist_model.py) for example)
-2. Write your PyTorch Lightning datamodule (see [mnist_datamodule.py](src/datamodules/mnist_datamodule.py) for example)
+1. Write your PyTorch Lightning model (see [models/mnist.py](src/models/mnist.py) for example)
+2. Write your PyTorch Lightning datamodule (see [datamodules/mnist.py](src/datamodules/mnist.py) for example)
 3. Write your experiment config, containing paths to your model and datamodule
 4. Run training with chosen experiment config: `python run.py experiment=experiment_name`
 
@@ -717,7 +717,7 @@ You can use many of them at once (see [configs/logger/many_loggers.yaml](configs
 
 You can also write your own logger.
 
-Lightning provides convenient method for logging custom metrics from inside LightningModule. Read the docs [here](https://pytorch-lightning.readthedocs.io/en/latest/extensions/logging.html#automatic-logging) or take a look at [MNIST example](src/models/mnist_model.py).
+Lightning provides convenient method for logging custom metrics from inside LightningModule. Read the docs [here](https://pytorch-lightning.readthedocs.io/en/latest/extensions/logging.html#automatic-logging) or take a look at [MNIST example](src/models/mnist.py).
 
 <br>
 
@@ -803,7 +803,7 @@ The following code is an example of loading model from checkpoint and running pr
 from PIL import Image
 from torchvision import transforms
 
-from src.models.mnist_model import MNISTLitModel
+from src.models.mnist import MNISTLitModel
 
 
 def predict():
@@ -880,7 +880,7 @@ You can easily modify the commands in the scripts for your use case. If 1 epoch 
 
 ### Callbacks
 
-Template contains example callbacks enabling better Weights&Biases integration, which you can use as a reference for writing your own callbacks (see [wandb_callbacks.py](src/callbacks/wandb_callbacks.py)).<br>
+Template contains example callbacks enabling better Weights&Biases integration, which you can use as a reference for writing your own callbacks (see [wandb.py](src/callbacks/wandb.py)).<br>
 To support reproducibility:
 
 - **WatchModel**
@@ -1324,8 +1324,8 @@ pip install git+git://github.com/YourGithubName/your-repo-name.git --upgrade
 So any file can be easily imported into any other file like so:
 
 ```python
-from project_name.models.mnist_model import MNISTLitModel
-from project_name.datamodules.mnist_datamodule import MNISTDataModule
+from project_name.models.mnist import MNISTLitModel
+from project_name.datamodules.mnist import MNISTDataModule
 ```
 
 </details>
