+++ configs/callbacks/wandb.yaml
@@ -2,25 +2,25 @@ defaults:
   - default.yaml
 
 watch_model:
-  _target_: src.callbacks.wandb_callbacks.WatchModel
+  _target_: src.callbacks.wandb.WatchModel
   log: "all"
   log_freq: 100
 
 upload_code_as_artifact:
-  _target_: src.callbacks.wandb_callbacks.UploadCodeAsArtifact
+  _target_: src.callbacks.wandb.UploadCodeAsArtifact
   code_dir: ${work_dir}/src
 
 upload_ckpts_as_artifact:
-  _target_: src.callbacks.wandb_callbacks.UploadCheckpointsAsArtifact
+  _target_: src.callbacks.wandb.UploadCheckpointsAsArtifact
   ckpt_dir: "checkpoints/"
   upload_best_only: True
 
 log_f1_precision_recall_heatmap:
-  _target_: src.callbacks.wandb_callbacks.LogF1PrecRecHeatmap
+  _target_: src.callbacks.wandb.LogF1PrecRecHeatmap
 
 log_confusion_matrix:
-  _target_: src.callbacks.wandb_callbacks.LogConfusionMatrix
+  _target_: src.callbacks.wandb.LogConfusionMatrix
 
 log_image_predictions:
-  _target_: src.callbacks.wandb_callbacks.LogImagePredictions
+  _target_: src.callbacks.wandb.LogImagePredictions
   num_samples: 8
