+++ run.py
@@ -13,7 +13,7 @@ def main(config: DictConfig):
     # Imports can be nested inside @hydra.main to optimize tab completion
     # https://github.com/facebookresearch/hydra/issues/934
     from src.train import train
-    from src.utils import utils
+    from src import utils
 
     # A couple of optional utilities:
     # - disabling python warnings
