+++ src/train.py
@@ -11,7 +11,7 @@ from pytorch_lightning import (
 )
 from pytorch_lightning.loggers import LightningLoggerBase
 
-from src.utils import utils
+from src import utils
 
 log = utils.get_logger(__name__)
